src/Orion_agent/requirements.txt
uagents>=0.4.0
requests>=2.28.0
numpy>=1.24.0
python-dotenv>=1.0.0